[qatl] moved exception to own package
+ added exceptions for auth API XMLRPC
+ removed old code
