+++ tests/unitaries/suite_001_connections.py
@@ -7,12 +7,9 @@ from unittest import TestCase
 from unittest import skip
 from qatestlink.core.testlink_manager import TLManager
 from qatestlink.core.models.tl_models import TProject
-# DONT DELETE THIS COMMENTS BEFORE REMOVE old_code directory ( just local )
-#from xml.etree.ElementTree import Element
-#from qatestlink.core.TlConnectionBase import TlConnectionBase
-#from qatestlink.core.xmls.XmlParserBase import XmlParserBase
-#from qatestlink.core.objects.TlTestProject import TlTestProject
-#from tests.config import Config
+from qatestlink.core.exceptions.response_exception import ResponseException
+
+
 API_DEV_KEY = 'ae2f4839476bea169f7461d74b0ed0ac'
 
 
@@ -66,6 +63,6 @@ class TestConnectionRaises(TestCase):
 
     def test_001_raises_connemptydevkey(self):
         """TODO: doc method"""
-        self.assertFalse(
-            self.testlink_manager.api_login(
-                dev_key=''))
+        self.assertRaises(
+            ResponseException,
+            self.testlink_manager.api_login, dev_key='')
